Fix for non linux OS build
